requirements-dev.txt
autopep8==1.3.5
pylint==1.9.2
pylint-django
pycodestyle==2.3.1
flake8
coverage

requests_mock
testing.redis~=1.1.0

ipdb
pdbpp

pytest
pytest-cov
pytest-django
mypy
